Fix source map
